Set tests up to run in Jenkins
