+++ package.json
@@ -17,9 +17,11 @@
     "angularjs-dropdown-multiselect": "^2.0.0-beta.10",
     "angularjs-slider": "~6.5.1",
     "formatcoords": "~1.1.3",
+    "grunt": "^1.0.3",
+    "grunt-cli": "^1.3.1",
+    "leaflet": "0.7.5",
     "leaflet-mouse-position": "1.0.1",
     "leaflet.nontiledlayer": "1.0.3",
-    "leaflet": "0.7.5",
     "mgrs": "~1.0.0"
   },
   "devDependencies": {
